AD: added modules format for fundamentals
