+++ src/main/resources/static/Index.html
@@ -9,6 +9,5 @@
 	</body>
 	<footer></footer>
 
-<script src="./scripts/vale-js-practice-script.js"></script>
-
+<script type="module" src="./scripts/vale-js-practice-script.js"></script>
 </html>
