+++ src/main/resources/static/scripts/fundamentals/util/vale-js-utilities.js
@@ -1,0 +1,12 @@
+"use strict";
+// ========================================
+// Section end
+
+const sectionEnd = () => {
+	console.log("\n========================================\nSECTION END\n========================================");
+}
+const sectionHead = (section) => {
+	console.log("\n" + section + "\n========================================");
+}
+
+export {sectionHead, sectionEnd};
